==> Chapter_4/4. QC.Rmd <==
---
title: "3. QC"
author: "Regan"
date: "07/04/2021"
output: html_document
---

```{r}
# set the directory
dirA <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 4/scRNAseq/"

# load the sce object
sce <- readRDS(paste0(dirA, "/1. QC/preQC_sce.RDS"))
sce$log10sum <- log10(sce$sum)
sce$log10detected <- log10(sce$detected)
```

#update sample names
```{r}
#save a copy of the old sample names
sce$sampleID -> sce$sampleID_old

#remove SIGA (common to all samples)
sce$sampleID <- gsub("SIGA", "", sce$sampleID)

sce$sampleID <- factor(sce$sampleID, levels = c("A10", "H2", "B11", "E5", "A1","F5","A7", "B7","C11","D11", "B10", "C10", "E3.D10", "F3.D10", "E5.D21", "F4", "F5.D21", "G4", "G5", "H5"))

table(sce$sampleID)
```

#update strain names
```{r}
sce$strain[sce$strain == "CreRFP"] <- "Resident"
sce$strain[sce$strain == "CreYFP"] <- "Infiltrating"

table(sce$strain)
```


#plot the QC metrics for all the samples
```{r}
dir0 <- paste0(dirA, "QC/")
dir.create(dir0)

stage = "PreQC, combined" #specify the stage for the plotting function
source("/Users/reganhamel/Documents/PhD/scRNAseq Workflow/PhD_Dissertation/Chapter_4/Sources/Fxn - qcPlots.R", echo = TRUE)
```



#plot the library size and complexity for each sample
```{r}
library(RColorBrewer)
library(ggplot2)
library(scater)

#set the dir
path <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 4/scRNAseq/1. QC/violin plots/"
dir.create(path)
y <- "log10sum"
c_by <- "strain" #"subsets_Mito_percent"
#to plot coldata
p <- plotColData(
  sce,
  y = y,
  x = "sampleID",
  colour_by = c_by
) +  theme(axis.text.x = element_text(angle=45, vjust = 1, hjust=1)) 

p
#save
ggsave(paste0(path, y,  "-", c_by, "-pre QC.png"), scale=1.5, height=2.5, width=8)
```

#plotting function
```{r}
library(ggplot2)

qcplot <- function (x, y, data, colour, col.high = "yellow", col.low = "blue", xlab = "x", ylab = "y", xlim =1.1*max(dat$x), ylim = 1.1*max(dat$y), size=I(0.3), col.label){ #list of arguments the fxn takes
  return (
    qplot(x, y,  data=dat, colour=colour, size=size) + scale_colour_gradient(low=col.low,high=col.high) + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank(), axis.line = element_line(colour = "black")) + scale_x_continuous(name=xlab, limits = c(0.99*min(dat$x), xlim)) + scale_y_continuous(name=ylab, limits = c(0.99*min(dat$y), ylim)) + labs(colour = col.label)
  )
}
```


```{r}
path <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 4/scRNAseq/QC/"

dat <- data.frame(x = log10(sce$sum), y = sce$subsets_Mito_percent, colour = sce$subsets_Mito_percent)
qcplot(dat$x, dat$y, data = dat, colour = sce$subsets_Mito_percent, xlab = "Log10 UMI Counts", ylab = "% Mito Counts", col.label = "% Mito Counts") + geom_hline(yintercept = 5, colour = "black")

ggsave(paste0(path,"mito_dot-pre QC.png"), scale=1)
```
# remove based on mito
```{r}
rem.mito <- sce$subsets_Mito_percent > 5
sce_mito <- sce[,rem.mito] #save the removed cells
sce <- sce[,!rem.mito]
```

#violin plots post-QC
```{r}
library(RColorBrewer)
library(ggplot2)
library(scater)

#set the dir
path <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 4/scRNAseq/QC/violin plots/mt_rem_"

y <- "log10detected"
c_by <- "strain"
#to plot coldata
p <- plotColData(
  sce,
  y = y,
  x = "sampleID",
  colour_by = c_by
) +  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) + geom_hline(yintercept = log10(600), colour = "black")

p
#save
ggsave(paste0(path,"violin-",y,  "-", c_by, "-pre QC.png"), scale=1.5, height=2.5, width =8)
```
# remove based on UMI/genes
```{r}
#for v3 chemistry
rem.sum <- sce$sum < 1000
rem.det <- sce$detected < 600
discard <- rem.sum | rem.det

sce_low <- sce[,discard]
sce <- sce[, !discard]
```

#violin plots post-QC (sanity check)
```{r}
y <- "log10sum"
sce$log10sum <- log10(sce$sum)
sce$log10detected <- log10(sce$detected)
c_by <- "subsets_Mito_percent"
#to plot coldata
p <- plotColData(
  sce,
  y = y,
  x = "sampleID",
  colour_by = c_by
) +  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) + geom_hline(yintercept = log10(600), colour = "black")

p
```

# check to see if a certain cell type is being discarded with these thresholds
```{r}
sce_mito2 <- sce_mito[,sce_mito$sampleID == "SIGAH5"]
sce_low2 <- sce_low[,sce_low$sampleID == "SIGAH5"]
discard <- cbind(sce_mito2, sce_low2)
remove <- calculateAverage(counts(discard))
keep <- calculateAverage(counts(sce))

library(edgeR)
logs <- cpm(cbind(remove, keep), log=T, prior.count=2)
logFC <- logs[,1] - logs[,2]
ave_count <- rowMeans(logs)

#highlight mito genes
is.mito <- which(rowData(sce)$CHR =="MT")
```

# colour the genes by cell type (kept)
```{r}
# contaminating cell types
#expected cell types
MG<- c("Olfml3","Sparc", "Trem2", "C1qa", "C1qb", "C1qc", "Ctss","Csf1r", "Hexb","Gpr34", "P2ry12","Plxdc2", "Sall1", "Siglech","Cx3cr1","Tmem119","Serpine2", "Cst3")
NP <- c("S100a8", "S100a9", "Lcn2", "Dedd2", "Ly6g")
MC <- c("Ccl7", "Ecm1", "Mgst1", "Arg1", "Ms4a7", "Fcer1g", "Gpnmb", "Ftl1", "Tyrobp", "Lgals3", "Lyz2", "Cyba", "Lgals3", "Tmsb4x", "Actb", "Fth1", "Ctsd", "Eef1a1", "Fau", "Rps8", "Rpl41", "Ctsb", "Rps29")
DC <- c("Cd74", "H2-Eb1", "H2-Aa","H2-b1", "H2-Ab1", "Flt3", "Zbtb46" )
aMG <- c("Spp1", "Fabp5", "Apoe")
gene <- list(Microglia=MG, Monocyte=MC, Neutrophils=NP, Dendritic_cells=DC, Activated_microglia=aMG)
col <- list(Microglia="red", Mono_macro="yellow", Neutrophils="purple", Dendritic_cells="orange", Activated_micro="green")

i = DC
#i = logFC > 0.25 & logFC < 0.4
jpeg(paste0(dirA, "QC/QC Genes/H5_kept.jpeg"), res=250, pointsize=6, width = 800, height = 800)
plot(ave_count, logFC, xlab="Average UMI count", ylab="logFC (removed/kept)", pch=18)
points(ave_count[MG], logFC[MG], col="red", pch=18)
points(ave_count[MC], logFC[MC], col="yellow", pch=18)
points(ave_count[NP], logFC[NP], col="purple", pch=18)
points(ave_count[DC], logFC[DC], col="orange", pch=18)
points(ave_count[aMG], logFC[aMG], col="green", pch=18)
legend(13.5, 1.7, legend=c(names(col)),
col=unlist(col), pch=18, cex=0.8)
#text(ave_count[i], logFC[i],  names(ave_count[i]), cex=0.65, pos=1,col="blue")
#points(ave_count[i], logFC[i], col="green", pch=18)

dev.off()
```

# colour the genes by cell type (discarded)
```{r}
# contaminating cell types
N <- c("Camk2n1", "Ckb", "Tubb2a", "Kif5b", "Tuba1a", "Chchd10", "Slc25a4", "Cpe", "Clu", "Nefl", "Camk2n1", "Pcsk1n", "Prnp", "Mt3", "Ap1s2", "Atp5k", "Hnrnpa2b1", "Hspa8")
AS <- c("Gfap", "Aldh1l1", "Slc1a3","Slc1a2", "Aqp4", "Atp1b2",  "Atp1a2","Aldoc", "Ntsr2", "Lcn2", "Ttyh1", "Chchd10", "Ndrg2", "Mobp", "Atp1b2", "Bcas1", "Sdc4", "Mal", "Gpr37l1")
OL <- c("Mbp", "Scd2", "Plp1", "Cnp", "Olig1", "Apod", "Plekhb1", "Cryab", "Cldn11", "Cldn11", "Gpm6b", "Ptgds")
EC <- c("Hspb1", "Sparcl1", "Sparc", "S100a16", "Isg15", "Ly6a")

gene <- list(Neurons=N, Astrocytes=AS, Oligodendrocytes=OL, Endothelial_cells=EC)
col <- list(Neurons="red", Astrocytes="yellow", Oligodendrocytes="purple", Endothelial_cells="orange", Mito_genes="blue")

jpeg(paste0(dirA, "QC/QC Genes/H5_All_types.jpeg"), res=250, pointsize=6, width = 800, height = 800,)
plot(ave_count, logFC, xlab="Average UMI count", ylab="logFC (removed/kept)", pch=18)
points(ave_count[N], logFC[N], col="red", pch=18)
points(ave_count[AS], logFC[AS], col="yellow", pch=18)
points(ave_count[OL], logFC[OL], col="purple", pch=18)
points(ave_count[EC], logFC[EC], col="orange", pch=18)
points(ave_count[is.mito], logFC[is.mito], col="blue", pch=18)
legend(13, 1, legend=c(names(col)),
       col=unlist(col), pch=18, cex=0.8)
#text(ave_count[i], logFC[i],  names(ave_count[i]), cex=0.65, pos=1,col="blue")
dev.off()
```

#remove sample SIGAF3 (88 cells)
```{r}
sce <- sce[,sce$sampleID != "SIGAF3"]
sce$sampleID <- droplevels(sce$sampleID)
```

#plot the QC metrics for all the samples
```{r}
stage = "PostQC, combined" #specify the stage for the plotting function
source("/Users/reganhamel/Documents/PhD/scRNAseq Workflow/PhD_Dissertation/Chapter_4/Sources/Fxn - qcPlots.R", echo = TRUE)
```

#identify doublets in the post-QC dataset
MUST BE DONE BY SAMPLE
```{r}
library(scds)
doublet <- NULL
doublet_score <- NULL
#identify doubles for each sample
for (i in levels(factor(sce$sampleID))) {
  SCE <- sce[,sce$sampleID == i]
  SCE <- cxds_bcds_hybrid(SCE, estNdbl=TRUE)
  doublet <- c(doublet, SCE$hybrid_call) #label now and remove later
  doublet_score <- c(doublet_score, SCE$hybrid_score) #label now and remove later
  
}

sce$doublet <- doublet
sce$doublet_score <- doublet_score

# Save the updated sce for each sample for downstream analysis
saveRDS(sce, paste0(dirA, "sce_dbl_marked.RDS"))
```

#plotting function - discrete colour values
```{r}
library(ggplot2)

qcplot <- function (x, y, data, colour, col.high = "yellow", col.low = "blue", xlab = "x", ylab = "y", xlim =1.1*max(dat$x), ylim = 1.1*max(dat$y), size=I(0.3), col.label){ #list of arguments the fxn takes
  return (
    qplot(x, y,  data=dat, colour=colour, size=size) + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank(), axis.line = element_line(colour = "black")) + scale_x_continuous(name=xlab, limits = c(0.99*min(dat$x), xlim)) + scale_y_continuous(name=ylab, limits = c(0.99*min(dat$y), ylim)) + labs(colour = col.label)
  )
}
```

#plot doublets
```{r}
path <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 4/scRNAseq/QC/"

dat <- data.frame(x = log10(sce$sum), y = log10(sce$detected), colour = sce$doublet)
qcplot(dat$x, dat$y, data = dat, colour = sce$doublet, xlab = "Log10 UMI counts", ylab = "Log10 detected genes", col.label = "Doublet score") + geom_hline(yintercept = 5, colour = "black")

#ggsave(paste0(path,"mito_dot-pre QC.png"), scale=1)
```

#remove doublets & failed samples
```{r}
sce <- sce[,!sce$doublet]
sce <- sce[,sce$sampleID != "SIGAC10" & sce$sampleID != "SIGAH5" ]
sce$sampleID <- droplevels(sce$sampleID)

saveRDS(sce, paste0(dirA, "sce.RDS"))
```

#record removed cells
```{r}
#print the cells removed
list <- list(
  table(discard),
  table(rem.mito),
  table(doublet),
  summary(discard | doublet))

capture.output(print(list), file = paste0(dir0, "/removed.csv"))
```

#median genes per cell
```{r}
med <- NULL
names <- NULL
for (i in levels(sce$sampleID)) {
  SCE <- sce[,sce$sampleID == i]
  med <- c(med, median(SCE$detected))
  names <- c(names, i)
}

names(med) <- names
med
```

#median/median genes per fate-map label
```{r}
lev <- sce$strain
med <- NULL
names <- NULL
for (i in levels(factor(lev))) {
  SCE <- sce[,lev == i]
  med <- c(med, mean(SCE$detected))
  names <- c(names, i)
}

names(med) <- names
med
```
