==> Chapter_4/5. CellAssign.Rmd <==
---
title: "Cell Assign"
---

```{r}
library(SingleCellExperiment)
library(stringr)

sce <- readRDS("/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 4/scRNAseq/2. BC/sce_UMAP_noBC.RDS")
dirA <- "/Users/reganhamel/Documents/PhD/PhD Dissertation/Chapter 4/scRNAseq/3. CellAssign/"
```


#set the marker lists for the cell types of interest
```{r}
marker_gene_list <- list(
  Microglia = c( "Sparc", "C1qa", "Plxdc2","Serpine2", "P2ry12", "Tmem119", "Siglech", "Ctss", "Cst3", "Slc2a5", "Sall1"), # "Sall1" not in HVG
  CAMs = c("Mrc1", "Lyve1", "Cd163", "Cbr2", "Ms4a7", "Cd74"), #low Siglec1 #no CAM genes in HVG
  Dendritic_Cells = c("Cd74", "H2-Eb1", "H2-Aa", "Plac8", "H2-Ab1", "H2-DMb1", "H2-DMa", "Klrd1", "Flt3", "Zbtb46"), #Itgax (expressed in activated MG too) # not in HVG
  Monocyte = c("Chil3", "Plac8", "Ccr2", "Rgs1", "Fn1"), #Vim (expressed in others)
  Macrophage = c("Ms4a7", "Ecm1", "Arg1", "Ccl7", "Mgst1"), 
  Neutrophil = c("S100a8", "S100a9", "Lcn2", "Dedd2", "Retnlg", "Wfdc21", "Mmp9"), #"Ly6g" not in HVG
 TCell = c("Cd2",   "Cd3d",  "Cd3e",  "Cd3g",  "Cd28",  "Cd4",   "Ptprc") #"Vim" #no TCell genes in hVGs
)


library(tensorflow)
library(cellassign)

markers <- c(marker_gene_list)#, pre_set)
marker_mat <- marker_list_to_mat(markers)

head(rownames(marker_mat))

#remove marker genes not in the HVG list
rem <- !(unlist(markers) %in% rownames(sce))
unlist(markers)[!(unlist(markers) %in% rownames(sce))]
```

#prepare the sce with the gene list
```{r}
#subset the sce to match the marker genes
sce_marker <- sce[rownames(marker_mat)]

#keep a record of the marker genes used
saveRDS(marker_gene_list, paste0(dirA, "/marker_gene_list.RDS"))
```


#predict the cell type
```{r}
fit <- cellassign(exprs_obj = sce_marker[rownames(marker_mat),], 
                  marker_gene_info = marker_mat, 
                  s = sizeFactors(sce_marker), 
                  shrinkage = TRUE,
                  verbose = F,
                  threads =4,
                  min_delta = 2)

saveRDS(fit, paste0(dirA, "fit.RDS"))
system("say assignment done")
```

#We can access the maximum likelihood estimates (MLE)  of cell type using the `celltypes` function:

```{r}
summary(factor(celltypes(fit, assign_prob = .80)))

```

# save the predicted cell type in the sce
```{r}
sce$cell_type_ca <- sce$cell_type <- celltypes(fit, assign_prob = .80)
```

```{r}
library(pheatmap)
pheatmap(cellprobs(fit), filename = paste0(dirA,"fit_plot.png"))

system("say plot done")
```

```{r}
saveRDS(sce, paste0(dirA, "sce.RDS"))
```


```{r}
library(scater)

cby <- "Cd2"
plotUMAP(sce, colour_by=cby)
#ggsave(filename = paste0(cby, ".png"), path = dirA, scale = 1, width = 8, height = 8,  dpi = 300, limitsize = TRUE)

 #TCell = c("Cd2",   "Cd3d",  "Cd3e",  "Cd3g",  "Cd28",  "Cd4",   "Ptprc") #"Vim" #no TCell genes in hVGs

```
#close look at cell types
```{r}
#rename levels to include NA for those not of interest
sce$grey <- sce$cell_type_ca 
#sce$grey[sce$date  != "1-May-21" & sce$date !=  "6-Mar-20" | (sce$stage !=  10 & sce$stage !=  21)] <- NA
sce$grey[sce$cell_type_ca != "CAMs"] <- NA

#plot
plotUMAP(sce, colour_by="grey")  + theme(plot.title = element_text(size = rel(2))) + scale_fill_manual(values = c("#D53E4F", "#542788", "#5AB4AC", "#A1D76A", "#F1A340", "#FFEDA0"), aesthetics = "fill", na.value = "grey90") 
```

#Cluster
```{r}
library(scran)
set.seed(100)

metadata(sce)$k <- 20
#generate the shared nearest neighbour graph
snn.gr <- buildSNNGraph(sce, k = metadata(sce)$k, type="jaccard") #k ranges from 5-100
#the larger k (# of nn), the fewer the clusters, 20 is the default in Seurat, 10 in scran
#lower k means the noise will have a greater impact on the clustering. Larger values are computationally expensive 

#apply the louvain commuinity detection
clusters <- igraph::cluster_louvain(snn.gr) 

#store the clusters in the sce object column-based metadata
sce$Cluster <- factor(clusters$membership)
table(sce$Cluster)
saveRDS(sce, paste0(dirA, "sce.RDS"))
```



#Manual touch-ups
```{r}
#investigate the cluster membership of unasigned cells
table(sce$strain, sce$cell_type)
```


#manually assign cell types to unassigned cells based on cluster
```{r}
list <- list(
  "1"="Microglia",
  "2"="Dendritic_Cells",
  "3"="Monocyte",
  "4"="Dendritic_Cells",
  "5"="Dendritic_Cells",
  "6"=NULL,
  "7"="Monocyte",
  "8"="Macrophage",
  "9"="Macrophage",
  "10"="Neutrophil",
  "11"="Neutrophil",
  "12"="Microglia",
  "13"="Macrophage",
  "14"=NULL,
  "15"=NULL,
  "16"='Dendritic_Cells',
  "17"= "Microglia",
  "18"="Microglia",
  "19"=NULL,
  "20"="Microglia",
  "21"="Microglia",
  "22"="Dendritic_Cells")

for (i in 1:length(list)) {
CT <- list[[i]]

sce$cell_type[sce$Cluster == i & (sce$cell_type_ca == "unassigned" | sce$cell_type_ca == "other")] <- CT

}
#sanity check
table(sce$Cluster, sce$cell_type_ca)
table(sce$Cluster, sce$cell_type)
```

#assign resident cells to MG
```{r}
sce$cell_type[sce$strain == "Resident" & sce$cell_type_ca == "unassigned"] <- "Microglia"

```

#check fate mapping labels and correct MG/MCd
```{r}
table(sce$strain, sce$cell_type)
```

#note a few (9) NPs in the MG cluster - correct these
```{r}
sce$cell_type[sce$strain == "Infiltrating" & sce$cell_type == "Microglia"] <- "Macrophage"
sce$cell_type[sce$strain == "Resident" & sce$cell_type == "Macrophage"] <- "Microglia"

table(sce$strain, sce$cell_type)
```

#annotate CAMs
```{r}
cd74 <- counts(sce)["Cd74",] > 0
mrc1 <- counts(sce)["Mrc1",] > 0 
ms4a7 <- counts(sce)["Ms4a7",] > 0 
CAMs <- (cd74 & mrc1 & ms4a7 & sce$strain == "Resident")

table(sce[,CAMs]$cell_type)

sce$cell_type[CAMs] <- "CAMs"
```

#implement siglich as MG marker
```{r}
table(sce$cell_type)
sig <- counts(sce)["Siglech",] > 0
sce$cell_type[sig] <- "Microglia"
```


#remove the na.s
```{r}
#check for NAs
sce$Cluster[is.na(sce$cell_type)]
```


```{r}
library(scater)
plotUMAP(sce, colour_by="cell_type")
ggsave(filename = "cell_type.png", path = dirA, scale = 1, width = 8, height = 8,  dpi = 300, limitsize = TRUE)

```




```{r}
saveRDS(sce, paste0(dirA, "sce.RDS"))
```
